requirements.txt
beautifulsoup4==4.12.2
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
google==3.0.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
protobuf==3.20.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.2.3
